Migration definition changed

+++ database/migrations/2021_05_10_153800_handsetmaster.php
@@ -4,31 +4,66 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateHandsetmasterTable extends Migration
-{
+return new class extends Migration {
+    /**
+     * Run the migrations.
+     */
     public function up()
     {
-        Schema::create('handsetmaster', function (Blueprint $table) {
+        if (!Schema::hasTable("handsetmaster")) {
+            Schema::create("handsetmaster", function (Blueprint $table) {
+                $table->increments("Id");
+                $table
+                    ->string("Name")
+                    ->nullable()
+                    ->default("NULL");
+                $table
+                    ->integer("MakeId", 11)
+                    ->nullable()
+                    ->default("NULL");
+                $table
+                    ->integer("ModelId", 11)
+                    ->nullable()
+                    ->default("NULL");
+                $table
+                    ->integer("ColorId", 11)
+                    ->nullable()
+                    ->default("NULL");
+                $table->tinyInteger("IsActive", 1);
+                $table
+                    ->datetime("CreatedDate")
+                    ->nullable()
+                    ->default("NULL");
+                $table
+                    ->string("CreatedBy", 250)
+                    ->nullable()
+                    ->default("NULL");
+                $table
+                    ->datetime("UpdatedDate")
+                    ->nullable()
+                    ->default("NULL");
+                $table
+                    ->string("UpdatedBy", 250)
+                    ->nullable()
+                    ->default("NULL");
+                $table->primary("Id");
 
-            $table->increments(Id);
-            $table->string('Name')->nullable()->default('NULL');
-            $table->integer('MakeId',11)->nullable()->default('NULL');
-            $table->integer('ModelId',11)->nullable()->default('NULL');
-            $table->integer('ColorId',11)->nullable()->default('NULL');
-            $table->tinyInteger('IsActive',1);
-            $table->datetime('CreatedDate')->nullable()->default('NULL');
-            $table->string('CreatedBy',250)->nullable()->default('NULL');
-            $table->datetime('UpdatedDate')->nullable()->default('NULL');
-            $table->string('UpdatedBy',250)->nullable()->default('NULL');
-            $table->primary('Id');
-
-            $table->foreign('CreatedBy')->references('UserName')->on('user');
-            $table->foreign('UpdatedBy')->references('UserName')->on('user');
-        });
+                $table
+                    ->foreign("CreatedBy")
+                    ->references("UserName")
+                    ->on("user");
+                $table
+                    ->foreign("UpdatedBy")
+                    ->references("UserName")
+                    ->on("user");
+            });
+        }
     }
 
     public function down()
     {
-        Schema::dropIfExists('handsetmaster');
+        if (Schema::hasTable("handsetmaster")) {
+            Schema::dropIfExists("handsetmaster");
+        }
     }
-}
+};
